Revert "Bugfix"
