+++ backend/app/api/routers/__init__.py
@@ -3,7 +3,7 @@ from fastapi import APIRouter
 from app.api.routers.posts import router as posts_router
 from app.api.routers.auth import router as auth_router
 from app.api.routers.matching import router as matching_router
-from app.api.routers.dm import router as dm_rouer
+
 #routerの登録
 router = APIRouter()
 router.include_router(posts_router, prefix="/posts")
